genAI-labs/workshop/requirements.txt
# Core Dependencies
boto3==1.34.0
botocore==1.34.0

# Claude Code / Anthropic
anthropic==0.7.8

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Web Framework (for API examples)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0

# AWS CDK (for infrastructure examples)
aws-cdk-lib==2.100.0
constructs>=10.0.0,<11.0.0

